Update: - unnecessary controllers + fix Route + Model - Game + addEntityFramework
