--- GamePortal/GamePortal.Web.Api/Controllers/Quoridor/RegUserController.cs
@@ -1,29 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Net.Http;
-using System.Web.Http;
-
-namespace GamePortal.Web.Api.Controllers.Quoridor
-{
-    [RoutePrefix("api/Quoridor/RegUser")]
-    public class RegUserController : ApiController
-    {
-        /*
-        This actions needs to be implemented:
-         Log out        
-         Delete account
-         Open settings panel
-         Change move timeout
-         Enable the function "random fences"
-         Open edit profile panel
-         Edit name
-         Change pasword
-         Edit email
-         Edit pawn avatar
-         View leaderboard
-         Select game type
-         */
-    }
-}
