+++ GamePortal/GamePortal.Web.Api/GamePortal.Web.Api.csproj
@@ -143,8 +143,6 @@
     <Compile Include="Controllers\Battleship\UserController.cs" />
     <Compile Include="Controllers\BlackJack\GameController.cs" />
     <Compile Include="Controllers\DotsAndBlocks\GameController.cs" />
-    <Compile Include="Controllers\Quoridor\AdminController.cs" />
-    <Compile Include="Controllers\Quoridor\RegUserController.cs" />
     <Compile Include="Controllers\Quoridor\UserController.cs" />
     <Compile Include="Controllers\Sudoku\GameController.cs" />
     <Compile Include="Controllers\Sapper\SapperGameController.cs" />
