+++ GamePortal/Igro.Quoridor.Data/packages.config
@@ -1,6 +1,4 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Bogus" version="29.0.2" targetFramework="net48" />
   <package id="EntityFramework" version="6.4.4" targetFramework="net48" />
-  <package id="Ninject" version="3.3.4" targetFramework="net48" />
 </packages>
