--- GamePortal/Igro.Quoridor.Logic/Models/AdminDTO.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-
-namespace GamePortal.Logic.Igro.Quoridor.Logic.Models
-{
-    public class AdminDTO : RegPlayerDTO
-    {
-    }
-}
