+++ GamePortal/Igro.Quoridor.Logic/Models/GameDTO.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Igro.Quoridor.Logic.Models
+{
+    public class GameDTO
+    {
+        public int Id { get; set; }
+        public string Status { get; set; }
+    }
+}
