--- GamePortal/Igro.Quoridor.Logic/Models/RegPlayerDTO.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Web;
-
-namespace GamePortal.Logic.Igro.Quoridor.Logic.Models
-{
-    public class RegPlayerDTO
-    {
-        public int Id { get; set; }
-        [Required]
-        public string UserName { get; set; }
-        [Required]
-        public string FirstName { get; set; }
-        [Required]
-        public string LastName { get; set; }
-        [Required]
-        public string Password { get; set; }
-        [Required]
-        public string Email { get; set; }
-        [Required]
-        public DateTime DateOfBirth { get; set; }
-        public string Avatar { get; set; }
-    }
-}
