+++ GamePortal/Igro.Quoridor.Logic/Models/UserDTO.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Web;
 
@@ -8,6 +9,18 @@ namespace GamePortal.Logic.Igro.Quoridor.Logic.Models
     public class UserDTO
     {
         public int Id { get; set; }
-
+        [Required]
+        public string UserName { get; set; }
+        [Required]
+        public string FirstName { get; set; }
+        [Required]
+        public string LastName { get; set; }
+        [Required]
+        public string Password { get; set; }
+        [Required]
+        public string Email { get; set; }
+        [Required]
+        public DateTime DateOfBirth { get; set; }
+        public string Avatar { get; set; }
     }
 }
